pyproject.toml
[tool.poetry]
name = "fastapi-start"
version = "0.1.0"
description = ""
authors = ["Wilfredo Huancollo <wilfredo_sj@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8.0"
email-validator = "^1.2.1"
emails = "^0.6"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
requests = "^2.28.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.961"
pytest = "^7.1.2"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
